2024.06.05 19:19 add mountain list and mapper

+++ src/main/java/site/dearmysanta/domain/mountain/Mountain.java
@@ -1,5 +1,6 @@
 package site.dearmysanta.domain.mountain;
 
+import org.json.JSONObject;
 import org.springframework.web.multipart.MultipartFile;
 
 import lombok.AllArgsConstructor;
@@ -9,6 +10,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.util.Arrays;
 import java.util.List;
 
 @Getter
@@ -26,7 +28,7 @@ public class Mountain {
     private String mountainLocation; // mountainLocationList.toString();
     private String mountainImage; // mountain image
     private String mountainDescription; // mountain description
-    private int mountainAltitude; // mountain altitude
+    private double mountainAltitude; // mountain altitude
     private int mountainTrailCount; // number of trails
     private int likeCount; // like count
     private int mountainViewCount; // view count
@@ -34,4 +36,6 @@ public class Mountain {
     private int calandarRegisteredCount; // calendar registered count
     private int meetingPostCount; // meeting post count
     private List<MountainTrail> mountainTrail; // list of mountain trails
+
+    
 }
